Extract HTTP request authentication to base class
